closes #86 adds in UI for recorder playback, and splits out user media from recorder
