+++ lib/rack/playback.mjs
@@ -1,60 +1,13 @@
+import {RackItem} from './RackItem.mjs';
 import {GLOBAL as GlobalScheduler} from '../scheduler/scheduler.mjs';
+import {PlaybackNode} from '../audio/playback-node.mjs';
 import {RingBuffer} from '../util/ringBuffer.mjs';
 
-export class PlaybackNode extends AudioWorkletNode {
-  
-  constructor(context, options) {
-    
-    super(context || GlobalScheduler.audioContext, 'playback-processor');
-    
-    this.channelCountMode = 'explicit';
-    this.channelInterpretation = 'speakers';
-    
-    this.port.onmessage = (e) => {
-      if (e?.data?.command === 'stop') {
-        this.dispatchEvent(new Event('ended'));
-      } else if (e?.data?.command === 'swap') {
-        const evt = new Event('swap');
-        evt.data = e.data;
-        this.dispatchEvent(evt);
-      }
-    };
-    
-    if (options?.sharedArrayBuffer) {
-      this.port.postMessage({
-        command: 'sab',
-        sharedArrayBuffer: options?.sharedArrayBuffer
-      });
-    }
-  }
-  
-  start(time) {
-    
-    const startTime = time || GlobalScheduler.audioContext.currentTime;
-    
-    this.port.postMessage({
-      command: 'start',
-      time: startTime
-    });
-  }
-  
-  stop(time) {
-    
-    const stopTime = time || GlobalScheduler.audioContext.currentTime;
-    
-    this.port.postMessage({
-      command: 'stop',
-      time: stopTime
-    });
-  }
-}
-
-export class Playback extends EventTarget {
+export class Playback extends RackItem {
   
   #buffer;
   #ringBuffer;
   
-  analyserNode;
   playbackNode = null;
   
   #handle;
@@ -62,7 +15,10 @@ export class Playback extends EventTarget {
   #streamHandler;
   
   constructor() {
-    super();
+    super({
+      enabled: false,
+      name: 'Playback'
+    });
     
     if (GlobalScheduler.audioContext) {
       this.#buffer = new SharedArrayBuffer(GlobalScheduler.audioContext.sampleRate * 3 * 2 * Float32Array.BYTES_PER_ELEMENT || 1000000);
@@ -74,17 +30,15 @@ export class Playback extends EventTarget {
     GlobalScheduler.addEventListener('AudioContextChanged', () => {
       this.#buffer = new SharedArrayBuffer(GlobalScheduler.audioContext.sampleRate * 3 * 2 * Float32Array.BYTES_PER_ELEMENT || 1000000);
     }, {once: true});
-  }
-  
-  init() {
     
     this.playbackNode = new PlaybackNode(GlobalScheduler.audioContext, {
       sharedArrayBuffer: this.#buffer
     });
     
-    this.analyserNode = new AnalyserNode(GlobalScheduler.audioContext);
-    
-    this.playbackNode.connect(this.analyserNode);
+    this.addOutput(this.playbackNode, {
+      type: 'Audio',
+      name: 'Main Out'
+    });
     
     this.playbackNode.addEventListener('swap', (e) => {
       if (e.data.high) {
@@ -97,7 +51,6 @@ export class Playback extends EventTarget {
         console.error(`swap buffer not specified`);
       }
     });
-    
   }
   
   set handle(handle) {
@@ -221,6 +174,10 @@ export class Playback extends EventTarget {
   
   start() {
     
+    if (!this.enabled) {
+      return;
+    }
+    
     if (!this.playbackNode || !this.playbackNode instanceof AudioNode) {
       throw new Error('call init before playing back');
     }
@@ -234,6 +191,3 @@ export class Playback extends EventTarget {
     this.#reader = null;
   }
 }
-
-await GlobalScheduler.audioContext.audioWorklet.addModule('lib/audio/playback-processor.mjs', {credentials: 'include'})
-  .catch((e) => {console.error(e, e.stack)});
